backend/requirements.txt
Django==2.2.8
django-debug-toolbar==2.0
psycopg2-binary>=2.8,<2.9
dj-database-url==0.4.1
ipython==5.3.0
ffmpeg-python==0.1.17
subliminal==2.1.0
gunicorn==19.9.0
clint==0.5.1 
django-cors-headers==3.0.2
djangorestframework==3.11.0
django-rest-auth==0.9.5
django-allauth==0.42.0
django-rest-auth[with_social]
django-redis==4.12.1 
pycodestyle==2.6.0
sentry-sdk==0.14.4
celery==4.4.6
redis==3.4.1
ffsubsync==0.4.3
baidupcs_py==0.7.1
celery-redbeat==2.0.0
aliyun-python-sdk-core
oss2
Levenshtein
srt
